refactor(notebook): extract environment lifecycle workflows
Move the five environment lifecycle commands behind an Electron-independent workflow while preserving the four IPC channels and Electron-only progress projection.
